backend/requirements.txt
# Core
python-dotenv==1.0.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# OpenAI - Older compatible version
openai==1.12.0

# LangChain - Compatible versions that work with openai==1.12.0
langchain-core==0.1.52
langchain==0.1.20
langchain-text-splitters==0.0.1
langchain-openai==0.0.8
langchain-community==0.0.38
langchain-pinecone==0.0.1

# Pinecone
pinecone-client==3.2.2

# Utilities
tiktoken==0.7.0
pandas==2.1.4
numpy==1.24.3

# API (for later)
fastapi==0.104.1
uvicorn==0.24.0